Restructure folder
1. Create module core with attackers and adversarial trainer
2. rename net to models with init
